Move modals from components to their own folder
